Watch nodes and link them to machines.
When the machine-controller starts up new hosts, those hosts already
self-annotate their corresponding node objects with the name of the
machine that was responsible for them. Now, a new watcher in the
machine-controller will look for these annotations and link the
Machine.Status.NodeRef to the proper node, and unlink them on node
deletion.