Use "new" machine object when deleting.
This fixes problems that arise when updating the Machine.Spec multiple
times, going from invalid to valid configuration. It's not a perfect
solution yet, so in the future we should probably start annotating
Machine objects with the resources that have been created for them, so
that we know the correct VM to delete if we need to replace it.