Correct the network configuration to allow the master to come up successfully (#114)
* Use kubenet instead of weave net

- add a worker tag to worker nodes so firewall rules created for
loadbalancers will be applied to the worker nodes.
- add the default service account to new VM instances.
- add serviceAccountActor role to the machine controller service account
so the master can spin up new nodes.
- add a function to the startup scripts to prevent docker from starting
on installation so we can configure DOCKER_OPTS and then start it
manually. This is needed to prevent it from messing with the IP tables.

* Remove --ignore-preflight-errors=all flag

The flag was replaced with a workaround to create the missing file that
triggered the error. The cluster was functioning when all preflight
errors were ignored, but in case another error comes up in the future,
we don't want to accidentally ignore it.

* Fix issue with assigning conflicting IPs to pods

* Revert "Kubenet"

* Update OS version to use ubuntu-1710-weave

This unbreaks the master branch where the network configuration isn't
working