Clean up adoption code, parallelize deletion. (#193)
This PR was split off from PR #165 to make it a smaller PR.

There is an edge case where the available status for a machine may be
miscalculated due to using the time from the node (last transition time)
and the current time of the controller. This may cause a non-trivial
delay during a deployment update due to an incorrect ready status when
it should be available.

Deletion calls are changed to be parallelized to enable speedier
reconcilation when machines are turned down.