Fix bug with deleting machines in non-default namespaces. (#189)
machineClient was created assuming namespace of default, thus an attempt
to delete a machine from another namespace would fail complaining that
the namespace on the request does not match the object. The machine
would then be effectively stuck, unable to clear it's finalizers.

This change instead uses the clientset itself to perform the update in
the correct namespace matching the Machine.