Revert changes to generated files (#231)
This will revert modifications to the generated files, and we work
around the initial reasons why we needed to modify them.

WaitForCacheSync can be called with a new channel, as it isn't really
used any more.

We store the shared informers for access to the queue instead of
storing the queue in Init.