Fix a race condition where machineset reconciles too quickly (#246)
The race condition is where the machineset reconciles on the same key
too quickly, where the creation/deletion of machines is not detected by
the second reconcilation, causing it create/delete additional machines.

I attempted to use WaitForCacheSync, but that is also insufficient in
preventing the race condition.

The fix here is to add 1 second sleep before releasing the mutex lock
when reconciling, which gives the system a chance to recognize the
changes made from the first reconciliation.

Issue #245 was created to improve this hacky fix.