Fix a bug in ClusterClient's GetClusterObjects() and GetMachineObjects() (#321)
The previous implementation took the address of the iteration variable. While the value of the iteration value changes on each iteration, the address does not change for the entirety of the loop. Previously, the slice returned by these functions would be a collection of the exact same iterator variable address and which of course would be filled with a value of the last value in the loop iteration.