Change the reconcile logic to include machines being deleted
This changes the logic from not counting machines that are being deleted
towards the machineset replica to counting them. The reason for the
change is that due to machine deletion taking non-trivial amount of
time, machine deployment rolling update may exceed the expected number
of machines utilized.

For example, for a machine set with 3 replicas and max surge 1, for a
max of 4 machines.
- create new machine in new machineset (4 machines)
- once new machine is ready, delete a machine from old machineset (4 machines, as machine is still being deleted)
- create new machine as deleted machine is not counted (5 machines)

For the duration of the old machine being deleted, we would exceed the
number of machines by 1.