Better way to register and lookup provisioners
Using a pattern similar to how we deal with cloud providers
in the main k/k repository.

Longer term plan is to fully eliminate ProviderDeployer interface, so
this registration mechanism is temporarily needed.

To enable this support in your provider, please add a init method
in your provider and use `RegisterClusterProvisioner` like so:

```
clustercommon.RegisterClusterProvisioner(ProviderName, actuator)
```

To use this you will need to edit clusterctl/main.go and add the package
that contains the init() method. You can just edit the copy of
clusterctl/main.go for trying things out during development or
vendor this file to your repo until we get rid of ProviderDeployer
and this registration mechanism.

Change-Id: Iaab03fccb32eed0cc6b4cccf46d50bda8fe2ab25
Signed-off-by: Davanum Srinivas <davanum@gmail.com>