Improve reliability of saving provider-components to configmap (#374)
This change increases the reliability of saving the provider components
to the cluster by moving the save to be below the apply of the
cluster API stack to the cluster. That apply process waits for the
cluster API server to be ready on some level so it improves the
reliability of saving the provider-components.

Ideally we should probably follow up with a wait process that ensures
the core kubernetes API server is ready to go.