Make MachineDeployment.Spec.Strategy a pointer (#551)
This is required in order to allow defaulting it via a MutatingWebhook.

When it is a literal and the user creates a MachineDeployment where that
property does not exist, the creation of a jsonpatch will happen based
on a representation where that field exists, since literals always gets
marshaled. This means a generated jsonpatch will get refused by the
APIServer which doesn't know anything about that field with an error
like this: `doc is missing path: "/spec/strategy/type"`