Add missing flags into cluster delete command (#561)
Added a few missing flags that were added to the create cluster
operation that would also be useful for delete cluster, such
as existing cluster, minikube options, and an option to clean
up the bootstrap cluster.  Also added a short hand for the
existing cluster for both create and delete cluster so users
don't have to type the whole flag name.

Also added a test for when delete cluster is called with no flags.
Currently, clusterctl will just hang.  Fixed that and added the
golden file for the test.

Fixes #560