Local defaulting (#599)
* Local defaulting

Until CRDs support defaulting, we can implement defaulting by simply
populating fields with their default values.

* Added PopulateDefaultsMachineDeployment to sync

With missing this, there are a couple of nil pointer panics, because the
content to wich `d` (or `deployment`) is pointing to, is changed before
storing it again. Subsequent code assumes `Strategy` to be there, no
matter wether or not it is already in the DB.

* Sync defaulting with old code

Syncing to f31486484d5b33c785540eeaffd47fe57832aef5

* Update bazel