Make APIVersion/Kind mandatory for MachineList
Reworked util.ParseMachineYaml to enforce APIVersion and Kind
as requirements to parse a MachineList object.

As this is a breaking change a new message is introduced when
parsing a YAML file that uses the old convention, test suite was
upgraded with a few cases.

Recommended to move to UniversialDeserializer once the API
is stabilized and added to k8s.io/client-go

Fixes: #717