Enable some commented out fields in the MachineClass type
This enables the `Capacity` and `Allocatable` fields on the
`MachineClass` type. I did a proof of concept[1] using the autoscaler
to see if adding these types would be sufficient to support
scale-from-0 - and they seemed sufficient for this use case which was
explicitly called out in the previous TODO comments.

[1] https://github.com/openshift/kubernetes-autoscaler/pull/89