Backports for release 0.1 (#924)
* Update CRDs (#863)

Generated with `make generate`

* fix various shell bugs (#868)

* fix errors in scripts/ci-build.gitsh

* fix errors in scripts/ci-integration.sh

* fix errors in scripts/ci-is-vendor-in-sync.sh

* fix errors in scripts/ci-make.sh

* fix errors in scripts/ci-test.sh

* removed unused variables from scripts/fetch_ext_bins.sh

* fix errors in hack/update-bazel.sh

* fix errors in hack/verify_clientset.sh

* use gnu sed on all platforms (#867)

* Fix error message in getCluster (#871)

Previously it was showing 'Machine foo in namespace "cluster.k8s.io/cluster-name" doesn't specify "k8s-machine-api"  label, assuming nil cluster'

* minikube: use --profile instead of --p (#875)

It was reported that "minikube start" fails:
  minikube start --bootstrapper=kubeadm --p=clusterapi
  Error: unknown flag: --p

From "minikube start --help":
  -p, --profile string

Use --profile in minikube.go as --p does not work.
The alternative is -p.

* nit: fix log to replace Creating with Preparing (#872)

saw this log with following flag
--bootstrap-cluster-kubeconfig /root/.kube/config

I0403 17:32:28.339021   31615 createbootstrapcluster.go:27] Creating bootstrap cluster

actually we are reusing ,not creating

* Record machine sync error event (#880)

* make clusterapi wait timeout configurable (#865)

add an environment to make timeout configurable
for timeoutMachineReady

* Update klog to 0.3.0 release (#890)

Signed-off-by: Vince Prignano <vincepri@vmware.com>

* Add missing namespace to Get function (#896)

* Add NodeName to Machine's additionalPrinterColumns (#895)

Because it is useful to cross-reference between Machines and Nodes.

Set priority to >0 so that node names will only be displayed in `-o wide` because the names are often quite long.

* additionalPrinterColumns for Machine, added correctly. (#897)

Co-authored-by: Danny Berger <dberger@pivotal.io>

* Normalize clusterctl options to underscores (#899)

Matches current conventions and implementation of kubectl.
This change does not include warning messages when using the previous,
underscore option names, but it does recognize those names.

Co-authored-by: Leah Hanson <lhanson@pivotal.io>

* Update manager-role ClusterRole (#901)

* Update Go 1.12 (#898)

* Update Go 1.12

Signed-off-by: Vince Prignano <vincepri@vmware.com>

* update

Signed-off-by: Vince Prignano <vincepri@vmware.com>

* Update OWNERS_ALIASES and SECURITY_CONTACTS (#884)

* Update OWNERS_ALIASES and SECURITY_CONTACTS

* Add davidewatson to cluster-api-admins

* Clarify log message for machine-controller isDeleteAllowed (#905)

Users have noted that this log message is confusing and
does not lead to a clear path forward when deleting this
machine.

This commit makes it clear that the machine will not be
deleted.

* Clarify MachineDeployment watchers' comments (#887)

Signed-off-by: Vince Prignano <vincepri@vmware.com>

* Simplify util.ParseMachinesYaml (#909)

Signed-off-by: Vince Prignano <vincepri@vmware.com>

* Add support for passing namespace cmdline flag to the manager (#914)

* Bump kubectl version to v1.13.2 in ci (#923)

kubectl v1.10.2 does not compatible with cluster v1.14.x anymore

Signed-off-by: Hui Luo <luoh@vmware.com>

* Add /status to MachineSet and MachineDeployment kubebuilder directives (#922)

Signed-off-by: Vince Prignano <vincepri@vmware.com>