UPSTREAM: <carry>: openshift: Ensure delete continues w/o node
This commit ensures we can continue with the instance
and machine deletion in the case that the node
referenced by the machine is already deleted.

Nodes can potentially be deleted by administrators
or possibly in previous Reconcile loop of the
machine-controller that failed to properly update
annotations.

This commit also adds logging to be more uniform
with how we log errors elsewhere in the controller.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1713105