Adds a failing test that should not fail
This test case shows a bug in the sync/rollout logic.

When the spec.template.spec of a deployment changes cluster-api does not
manage the machinesets correctly.

If there is a machinedeployment with 3 replicas and the spec defines a kubelet version
then the kubelet version is changed, the machinedeployment will spawn a new machineset
that has 1 replica with the new version and keep the old machineset around forever so
now you have 4 nodes.

Signed-off-by: Chuck Ha <chuckh@vmware.com>