Do not delete NodeRef for last control plane node (#1110)
* Do not delete NodeRef for last control plane node

This patch fixes a bug that occurs when the NodeRef is deleted for the
last control plane node in a cluster. A timeout error would occur
because the last machine would have already been deleted, and the
cluster would no longer be active to receive the delete node request.

* Addressing PR feedback

This patch addresses PR feedback:

* Use helper functions for constructing ListOptions
* Do not log unnecessary message regarding missing label
* Do not check for NotFound error

* Correct error message case

This patch corrects the case of an error message.

* Logging reason why NodeRef can't be deleted

This patch refactors the isDeleteNodeAllowed function a bit to increase
the logging around the reason the NodeRef could not be deleted.

* Addressing PR feedback

This commit addresses PR feedback:

* Changes log level of isDeleteNodeAllowed error
* Removes an extra space