Add section on conventions in Cluster API (#1213)
Getting started on the Cluster API projects, I found that the number
of acronyms can be confusing (e.g. CAPA: is it AWS or Azure?
CAPI: is it the API or the IBM Cloud provider?). This change aims to
help future newcomers to get familiar with the acronyms/abbreviations
used in the projects.

This change further makes small improvements to the tooling around
the GitBook: (1) ignore generated book files, and (2) add book build
and cleanup commands to the main Makefile.

Signed-off-by: erwinvaneyk <erwinvaneyk@gmail.com>