Update API endpoints if missing
This patch updates the behavior of the cluster controller's
"reconcileInfrastructure" function so that the API endpoints are always
reconciled if they have not been set and the infrastructure is supposed
to be otherwise ready.

Fix InfraReady/APIEndpoint order

This patch addresses the failure of the previous fix. This patch ensures
that "Cluster.Status.InfrastructureReady" is set to true even if API
Endpoints are not yet ready. However, the original intent of this patch
is retained -- API endpoints have more than a single opportunity to be
reconciled.