Fix MachineDeployment test flake
There was a race condition where a Machine was deleted from the
apiserver but the manager's shared informer cache hadn't caught up yet.
We were trying to update all the infra resources for all Machines,
including the Machines that had been deleted (but as far as the cache
was concerned, they hadn't been deleted yet).

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>