manager: use our RESTMapper
Switch to using our caching RESTMapper in the manager. If one of our
controllers encounters a Kind that was added to the apiserver after the
manager retrieved and cached the API discovery document, it won't be
able to work with it because this new Kind is not in the cache. We need
to invalidate and flush the cache in this scenario.

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>