:bug: refresh token for provisioning machines
Includes:

Configurable token duration to extend the required sync interval when
necessary.

Also adds a requeueAfter parameter, because I realized doing this work
that we should get an immediate second crack at the config before it's
consumed.

Finally, reconciles status when the config has already been used. In
order for us to get to that point the owner ref must be set by the
machine controller, which implies that the machine has a pointer to this
config object. So either it's an extremely unlikely user error, or we
mistakenly dropped the "ready" flag for this config.