Adds a tool to update certmanager YAML
Automating the pulling in of cert-manager.yaml through the hack/tools pattern.

We will likely be updating the cert-manager moving forward as controller-runtime updates
their version of cert-manager. However, they do not expose the version they support in an
obvious or programmatic way, so this is an attempt to make the manual step of updating
cert-manager as easy as possible.

Signed-off-by: Chuck Ha <chuckh@vmware.com>