Set cluster status phase to Provisioned after APIEndpoints are available.
Update cluster status phase to Provisioned after InfrastructureReady is true
and APIEndpoints are available. InfrastructureReady is set by provider
implementations to indicate that the provider is ready to be used and setting
cluster status to Provisioned just based on infrastructureReady can be a confusing
user experience.

Add unit tests for reconcilePhase.

Fixes: 1722