Reenable building & testing CAPD
capd: remove prereqs from test target

This is consistent with the main Makefile - we don't need to regenerate
and relint every time we call 'make test'.

capd: use preserveUnknownFields for CRD generation

With the upgrade to the newer version of controller-tools, we need to
specify crd:preserveUnknownFields to controller-gen when generating
CRDs. This aligns with a change already made to CAPI proper.

capd: enable linting & fix lint issues

Don't verify modules twice

For our CI tests, build/test capd as part of the main capi CI.

capd: fix verify boilerplate

It was cd'ing to the git repo root instead of
test/infrastructure/docker. We may want to consider removing the cap
copy of this, as it's covered by the core capi verification.

capd: remove verify-doctoc

capd: use main shellcheck

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>