Fix e2e framework resource apply order
This patch fixes the order in which resources are applied in the e2e
framework. Prior to this patch the cluster and infra cluster resources
were applied, and then the framework waited for the cluster to be
provisioned. The provision phase is not set by the cluster controller
until there is a non-zero control plane endpoint. For providers that get
the control plane endpoint from a machine, this meant that the cluster
would never be provisioned until machine resources existed.

This change is also consistent with how CAPI v1alpha2 works -- all
resources are applied in sequence before any action is taken to
determine the status of the deployed resources.