Add an option to specify the kind cluster name when using Tilt
Before this change, Tilt would attempt to preload images to a kind
cluster named `kind`.

With this change, a developer can use a kind cluster called something
other than `kind` by specifying the cluster name in the
tilt-settings.json file.

This commit also updates the docs to include a reference to the new
setting.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>