Fix sample Machine and MachineDeployment objects in Quick Start doc
The Quick Start documentation includes sample YAML definitions for
Machine and MachineDeployments. The spec of both these resources
requires a field called `clusterName`, which is missing from the sample
YAML.

Add the `clusterName` field to all sample Machine and MachineDeployment
resources that exist in the Quick Start section of the Book.

Signed-off-by: Alexander Brand <alexbrand09@gmail.com>