update worker kubernetes version in capd test template
This change removes the hardcoded version for a worker node and replaces
it with the same template variable that is used for the master. If used
as a template it would always deploy a worker at version "1.14.2", it
should now properly substitute the same version for both master and
worker nodes.