Add CAEP Extensible Template Processing for clusterctl
Resolved Open Questions:
- After investigation, we can use the clusterctl config file to define
nested values. Nested values such as
aws:
  secretKey: foo
will be read in as 'aws.secretkey'

- Infrastructure components yaml will be processed using the default yaml
processor. That is, 'clusterctl init' will not be accepting the
'--experimental-template-engine` flag as part of the initial work.

- We will call the existing variable substitution method as the
SimpleYamlProcessor. This will be the default mode of operation for
clusterctl.

- Remove references to specific template engine
This proposal is meant to provide an interface for any templating tool
that providers and operators may want to use. No need to mention any
specific templating tools

- Remove references to experimental flag UX with clusterctl since we are
only making library changes to expose the yaml processor usage.