⚠️ MachineHealthCheck Spec.Selector field cannot be empty
An empty selector is usually a valid selector that matches all objects;
in Cluster API usually we disallow its use given the scope of our
project is to manage the lifecycle of Kubernetes clusters and
having an "everything" selector could lead users into errors.

For MachineHealthCheck in particular, an empty selector would match
all the machines in a given cluster; given that we don't want to have
overlapping health checks, it makes sense to disable it and make it
opt-in.

Signed-off-by: Vince Prignano <vincepri@vmware.com>