:sparkles: KCP adopts existing machines
The KCP controller identifies Machines that belong to the control plane
of an existing cluster and adopts them, including finding PKI materials
that may be owned by the machine's bootstrap config and pivoting their
ownership to the KCP as well.

Prior to adopting machines (which, if unsuccessful, will block the KCP
from taking any management actions), it runs a number of safety checks
including:

- Ensuring the KCP has not been deleted (to prevent re-adoption of
  orphans, though this process races with the garbage collector)
- Checking that the machine's bootstrap provider was KubeadmConfig
- Verifying that the Machine is no further than one minor version off of
  the KCP's spec

Additionally, we set set a "best guess" value for the
kubeadm.controlplane.cluster.x-k8s.io/hash on the adopted machine as if
it were generated by a KCP in the past. The intent is that a KCP will
adopt machines matching its "spec" (to the best of its ability) without
modification, which in practice works well for adopting machines with
the same spec'd version.

We make an educated guess at what parts of the kubeadm config the operator
considers important, and feed that into the hash function.

The goal here is to be able to:
1. Create a KCP over existing machines with ~the same configuration (for
   then new api/equality's package definition of the same) and have the
   KCP make no changes
2. Create a KCP over existing machines with a different configuration
   and have the KCP upgrade those machines to the new config

Finally, replaces PointsTo and introduces IsControlledBy, both of which
search through the owner reference list of their first argument (similar to
metav1.IsControlledBy).

Unlike the metav1 function, they're looking for a match on name plus
api group (not version) and kind. PointsTo returns true if any
OwnerReference matches the pointee, whereas IsControlledBy only returns
true if the sole permitted (by the API) controller reference matches.

fix: handle JoinConfiguration hashing

refactor: util should not check UIDs

This is a behavior change not least for tests, which typically do not
set either Kind or GroupVersion on various objects that end up acting as
referents.

Co-Authored-By: mnguyen <mnguyen@newrelic.com>
Co-Authored-By: Jason DeTiberus <detiberusj@vmware.com>
Co-authored-by: Vince Prignano <vince@vincepri.com>