Fix for watch info matching in cluster cache tracker
The cluster cache tracker tries to deduplicate watch requests by matching the watchinput kind, watcher and eventhandler to the watches that it has already started.

However, this comparison was failing for real world event handlers because Go was not matching the inputs. Switching to reflect.DeepEqual helps with the equality of the watcher and kind, but still fails for eventHandlers because they will embed functions and DeepEqual will always return false for non-nil functions.

As a solution/workaround, we replace the eventHandler with a string representation of itself which will identify the struct type and should give pointers to the functions embedded within it. I'm hoping this should be sufficiently unique to identify a particular codepath calling Watch, when combined with the watcher.