Adds a test for ensureNamespace and fixes scope bug
This PR was initially to write a test for the `ensureNamespaces` function,
not to be confused with the `ensureNamespace` test I have in this commit.

Upon testing `ensureNamespaces`, I think I've come across a few bugs. This
function itself makes a call to `ensureNamespace` and it seems like there
is potentially a scoping issue with where `err` is being set.

It is my understanding that the following pattern:

if err := something(); err != nil {

}

implies that this particular error is available within the scope locally.
You can test out the differece by reverting the small change made in
`mover.go` to see the different in behaviour.

`ensureNamespaces` still needs a test written.