:bug: KCP reconcileDelete should not use RequeueAfterError
While looking for RequeueAfterError usage, to see how we can deprecate
in a few places, stumbled upon this usage in KCP which is incorrect.

Returning a RequeueAfterError here has no effect because the caller does
not look at the error result, but instead reconciliation will go in
a backoff loop.

Signed-off-by: Vince Prignano <vincepri@vmware.com>