Add clusterctl version check
- Store release info in state file
Cache the release info in the state file in order to avoid excessive
github requests.
- Use GITHUB_TOKEN for github client if available
- Handle special cases of GitVersions
We want to treat pre-releases as valid.
We log out if the cli version is a dev build of any kind
- Do not return error if unable to get github release
This can be the case for air-gapped environments. So we just log the error instead.
- Prompt upgrade only within minor versions of cli