:seedling: Save a DeepCopy when patching typed objects
When using the patch helper to patch objects, we perform lots of
DeepCopy for safety purposes. After giving a closer look, it seems we
can definitely save one deep copy when the patch helper is created and
when Patch() is invoked by checking if the incoming object is
unstructured or not.

The deep copy was previously performed because the unstructured coverter
in the runtime package returns the inner map of an unstructured object
instead of making it a copy (this isn't a bug, it's how it's meant to be
used).

Signed-off-by: Vince Prignano <vincepri@vmware.com>