:seedling: Save a DeepCopy when calculating patches in Patch helper
This PR reshuffles the code in a way that avoids copying the
unstructured objects first and then removing un-needed parameters from
the final object.

With these changes we opt to operate a shallow copy of the unstructured
objects with only the keys that are of interest, and then covert these
back to the actual object.

Signed-off-by: Vince Prignano <vincepri@vmware.com>