feat: add EKS controlplane provider
Add the EKS control plane provider to `clusterctl` which will be
available in the **0.6.1** release of CAPA.

The EKS bootstrap and control plane providers have had their names
changes so they are prefixed with aws and both end up being
`aws-eks`.

The apidiff test fails will fail with this changed due to the removal
of the  `EKSBootstrapProviderName` provider name.