Update linkchecker to use retry with exponential backoff
Turns out we can't use an API token for getting around the rate-limiting
of GitHub. We will try using retry with backoff, since each run of the
linkcheck writes to a cache and so will not re-check already successful
links.

Do not install linkchecker for serve/build

Marking plugins as optional means that they only run if installed, and
do not fail if not installed. We want to not install the linkchecker for
any targets except "verify" so that we don't run the verification in CI
environments that aren't specifically trying to do verification.

Remove it so that it works the same in dev environments where the
linkchecker may have been previously installed by running `verify`.