:warning: Use client.Object instead of runtime.Object
With controller runtime v0.7, most of the interfaces and clients now
accept client.Object or client.ObjectList instead of plain
runtime.Object. The original reason to have this change is to catch
errors at compile time rather than runtime. It also allows us to make
sure all objects we're dealing with have an object meta, which was
required before as well. The functionality is the same, the checks were
performed internally before while now the interfaces are compiled with
the expected objects.

Signed-off-by: Vince Prignano <vincepri@vmware.com>