:book: Relax our backporting guidelines and policies
This change relaxes and defines a bit more what can be backported and
how. After this change is merged, we'll allow both bug fixes and minor
(non-breaking) features to be backported if at least one or more
maintainers approves the cherry pick.

With these changes we aim to be more welcoming of changes to the current
stable branch instead of stopping the world until the next releases.

The only caveat is that backports that I'd like to highlight is that
backporting is an expensive operation. More often that not,
a cherry-picked commit requires another full review because changes are
usually not compatible.

Going forward, we won't allow PRs to target a release branch directly,
we'll always require changes to go in the main branch first, and only
after they've been merged, to backport them.

Signed-off-by: Vince Prignano <vincepri@vmware.com>