:seedling: Increase leader election lease values
To improve a self-managed cluster resilience to temporary errors related
to etcd leadership, this change increases the duration for all the lease
times.

The following are the most important values for leader election, we
increase the amount the non-leader candidates wait (1m now) and we
increase the renew deadline to 40s instead of 10, which should give
enough time for etcd connectivity to be established again.

- Lease duration is now 1 minute instead of 15s
- Renew deadline has been increased to 40 seconds instead of 10

In addition:

- Retry period has been increased to 5 seconds instead of 2
  - Avoid overloading the API Server / etcd with lease retry requests

Signed-off-by: Vince Prignano <vincepri@vmware.com>