Use uncached client and partial metadata for secret and configmaps
These changes allow the current stable Cluster API release to reduce its
memory footprint by a large margin.

Currently, we have multiple controllers that are either watching (like
ClusterResourceSet) or querying (get/list) corev1.Secret and
corev1.ConfigMap resources.

When these kinds go through the case, all of the objects in the cluster
end up being cached, not just the ones we're interested in.

In production environments, there might be a large number of ConfigMap
or Secret resources that we end up caching and watching for little gain.

With this patch, we introduce a new utility function that can generate
a delegating client that's smart enough to accept a list of objects that
we should always avoid caching

Signed-off-by: Vince Prignano <vincepri@vmware.com>