Uses patchHelper to update kubeadmConfig
Instead of issuing multiple update calls during the KCP upgrade, the
reconciler makes the changes to the same kubeadmConfig object and the
changes to the object are persisted using a patch() call.
This also removes the intermediate methods on the WorkloadCluster
interface and the reconciler calls these methods directly on the
kubeadmConfig object.
This also exposes a getter for workload cluster's client

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>