clusterctl completion zsh: Use native zsh completion
Due to this change, we will not able to load the zsh completion code
with the following method:
```
source <(clusterctl completion zsh)
```

Instead, we execute the following command once:
```
clusterctl completion zsh >"${fpath[1]}/_clusterctl"
```
Then we need to start a new shell for this setup to take effect.

---

In addition, github.com/spf13/cobra is updated to v1.1.1 in order to use
native zsh completion feature.