Updates controller-runtime version to v0.8.1
This involved bumping the code-gen version to 0.20 because the signature
of the function had been updated. Required regenerating the conversion
code.
Also updated the types containing slice of pointers of custom object
types to slice of custom objects to avoid conversion errors.
Update the v1a3 -> v1a4 upgrade docs to record this requirement.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>