E2E test to upgrade workload cluster
This patch adds an E2E test to create a workload cluster containing of
machinedeployment and machinepool. This cluster is then upgraded to
another k8s version, including the CP, machine pool and machine
deployment.
To verify the status of the upgrade, conformance tests are run on the
cluster post upgrade.

This also updates the e2e run target to include an additional parameter
named GINKGO_SKIP which can be used to skip ginkgo tests matching the
regexp.

Signed-off-by: Sagar Muchhal <muchhals@vmware.com>
Co-authored-by: Cecile Robert-Michon <cerobert@microsoft.com>
Co-authored-by: Lubomir I. Ivanov <neolit123@gmail.com>
Co-authored-by: Fabrizio Pandini <fabrizio.pandini@gmail.com>