:seedling: Remove verbose log line from CRS controller
This log line was logged every time the function ApplyClusterResourceSet
was called. This function is called every time CRS reconciles, even if
it doesn't have any more work to do, causing a lot of similar error
lines to be printed to users.

This change removes the log line for quieter logs.

Signed-off-by: Vince Prignano <vincepri@vmware.com>